Rename 'engine' package to 'graphics'
