+++ src/ru/nsu/ccfit/zuev/osu/LibraryManager.java
@@ -1,7 +1,7 @@
 package ru.nsu.ccfit.zuev.osu;
 
 import android.os.Build;
-import com.reco1l.osu.engine.VideoTexture;
+import com.reco1l.osu.graphics.VideoTexture;
 import com.rian.osu.beatmap.parser.BeatmapParser;
 import org.anddev.andengine.util.Debug;
 import org.jetbrains.annotations.Nullable;
