+++ src/ru/nsu/ccfit/zuev/osu/ResourceManager.java
@@ -9,7 +9,7 @@ import com.dgsrz.bancho.security.SecurityUtils;
 
 import com.reco1l.framework.IniReader;
 import com.reco1l.osu.conversion.SkinConverter;
-import com.reco1l.osu.engine.BlankTextureRegion;
+import com.reco1l.osu.graphics.BlankTextureRegion;
 import org.anddev.andengine.engine.Engine;
 import org.anddev.andengine.opengl.font.Font;
 import org.anddev.andengine.opengl.font.FontFactory;
