+++ src/ru/nsu/ccfit/zuev/osu/game/GameScene.java
@@ -13,8 +13,8 @@ import com.edlplan.osu.support.timing.TimingPoints;
 import com.edlplan.osu.support.timing.controlpoint.ControlPoints;
 import com.reco1l.api.ibancho.RoomAPI;
 import com.reco1l.osu.Execution;
-import com.reco1l.osu.engine.BlankTextureRegion;
-import com.reco1l.osu.engine.VideoSprite;
+import com.reco1l.osu.graphics.BlankTextureRegion;
+import com.reco1l.osu.graphics.VideoSprite;
 import com.reco1l.osu.ui.entity.InGameLeaderboard;
 import com.reco1l.osu.Multiplayer;
 import com.reco1l.osu.ui.multiplayer.RoomScene;
